// apps/web/src/components/PhoneInput.d.ts
export declare const COUNTRY_CODES: {
    code: string;
    flag: string;
    label: string;
}[];
interface PhoneInputProps {
    value: string;
    onChange: (value: string) => void;
    label: string;
    error?: string;
}
export declare function PhoneInput({ value, onChange, label, error }: PhoneInputProps): import("react/jsx-runtime").JSX.Element;
export {};
//# sourceMappingURL=PhoneInput.d.ts.map